backend_python_service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astroid==4.0.2
bcrypt==4.0.1
beanie==1.26.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
lazy-model==0.2.0
librt==0.7.3
mccabe==0.7.0
motor==2.5.1
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.4
pymongo==3.13.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
starlette==0.50.0
toml==0.10.2
tomlkit==0.13.3
types-passlib==1.7.7.20250602
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
